---sku信息SkuInfoController(分页获取sku信息)接口--- ---spu信息SpuInfoController(新增商品、多表同步操作)接口--- 时间处理MyMetaObjectHandler.class

+++ gulimall-product/src/main/java/com/hngc/product/entity/SpuInfo.java
@@ -45,7 +45,7 @@ public class SpuInfo implements Serializable {
     private BigDecimal weight;
 
     @ApiModelProperty("上架状态[0 - 下架，1 - 上架]")
-    private Byte publishStatus;
+    private Integer publishStatus;
     @ApiModelProperty("创建时间")
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createTime;
